+++ CHANGES.rst
@@ -4,8 +4,9 @@ Changes
 1.0.5 (unreleased)
 ------------------
 
-- Extend ``yafowil.referencebrowser`` by ``register_array_hooks``, ``array_add``,
-  ``array_index`` methods to enable usage in ``yafowil.widget.array``.
+- Extend ``yafowil.referencebrowser`` JS by ``register_array_hooks``,
+  ``array_add``, ``array_index`` methods to enable usage in
+  ``yafowil.widget.array``.
   [lenadax]
 
 - Use ``get_root`` in ``register_config`` and ``register_entry``.